Refactored so that XmlDocument is the only class that knows about Oga
The idea of XmlDocument was that it should encapsulate all xml access
methods (so that we could change the xml gem used or provide
adapters). However, DiffBuilder was also using oga, "leaking" the
encapsulation of oga.

To improve that I've refactored XmlDocument so that it provides a sax
parsing interface, which is now used by DiffBuilder, and DiffBuilder
no longer knows about the xml parsing library being used.

Going forwards all xml handling should be performed by the XmlDocument
class.