Made sure all internal methods receive XmlDocuments (not strings)
Previously, `Applicator.new` and `DiffBuilder.parse` would receive a
string which they would immediately parse into an
XmlDocument. However, this tightly couples them to how we parse
xml.

To reduce this coupling I've made both receive an XmlDocument object,
in the expectation that the caller should do the parsing for them.