Don't parse nested <remove> tags
There's a bug in the code at present where we will parse and apply
<remove> tags anywhere in the patch document, even deeply nested
ones. Instead we should only be parsing operation tags at /diff/*.

To make this work I've changed from using a SAX parser to a DOM
parser, as I couldn't see a way to either limit the depth of the oga
SAX parser or check a node's position in the document tree. Instead we
only get nodes at the specific xpath /diff/*